requirements.txt
ipython==8.12.3
langchain==0.3.26
langchain_community==0.3.27
langchain_openai==0.3.27
langgraph==0.5.0
Pillow==11.3.0
pydantic==2.11.7
selenium==4.34.0
torch==2.7.0
transformers==4.51.3
openai>=1.12,<2